pyproject.toml
[project]
name = "alexa-w"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "aiohttp==3.12.15",
    "grpcio==1.74.0",
    "grpcio-health-checking==1.74.0",
    "grpcio-tools==1.74.0",
    "numpy==2.2.6",
    "nvidia-ml-py3>=7.352.0",
    "onnxruntime==1.22.1",
    "openai-whisper==20250625",
    "openwakeword==0.6.0",
    "psutil==7.0.0",
    "pyyaml==6.0.2",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "sounddevice==0.5.2",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "tqdm==4.67.1",
    "webrtcvad==2.0.10",
]
